Makes env a const instead of define
makes delete_file fail safely and quietly

+++ src/file/file.cpp
@@ -189,6 +189,10 @@ void FileManager::delete_dir(string path) {
 }
 
 void FileManager::delete_file(std::string path) {
+  if(is_dir(path)) {
+    return;
+  }
+  
   unlink(path.c_str());
 }
